# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Fooseball api Documentation",
        "description": "The endpoints to connect to the fooseball api",
        "contact": {
            "name": "Van de Voorde Siebe",
            "email": "siebe.vandevoorde@student.odisee.be"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "api/teams": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get teams excluding the authenticated player",
                "description": "Display a listing of the resource.",
                "operationId": "237ce3475ae1e5a555a27effd9fbf82b",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Team1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Create a new team",
                "description": "Store a newly created resource in storage.",
                "operationId": "d6ef612b57e285ea01a58d3d9cfc4108",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "description": "The optional ID of the second player",
                                        "type": "integer",
                                        "format": "int64",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Team was made successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "api/teams/my": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get teams of the authenticated player",
                "description": "Display loggedin user teams.",
                "operationId": "2e7ba0b5d5d8e40f1e9e68af6644321d",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Team1"
                                                    },
                                                    "player1_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "player2_id": {
                                                        "type": "integer",
                                                        "example": 7
                                                    },
                                                    "total_wins": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "games_played": {
                                                        "type": "integer",
                                                        "example": 9
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}